traffic-counter-mvp/requirements.txt
ultralytics>=8.0.0
opencv-python
pandas
pyyaml
tqdm
